+++ NEWS.md
@@ -14,6 +14,7 @@
 
 * `var_names()` now also cleans variable names from variables modelled with the `offset()` function.
 * `icc()`, `re_var()` and `get_re_var()` now support `brmsfit`-objects (models fitted with the *brms*-package).
+* For `fun = "weighted.mean"`, `typical_value()` now checks if vector of weights is of same length as `x`.
 
 # sjstats 0.12.0